cmtool_core::grid

Struct BaseCompartmentMesh

pub struct BaseCompartmentMesh<T> {
    axes: [CoordAxis; 3],
    n_cells: usize,
    _marker: PhantomData<T>,
}

Base Compartment mesh structure for spatial modeling.

This struct provides the base components data for compartmental meshes

Type Parameters

T: A marker type used to distinguish different mesh configurations

Fields
axes: [CoordAxis; 3]

The axes of the coordinate system for the mesh.

This array contains three CoordAxis instances, each representing one of the principal axes

n_cells: usize

The total number of cells in the mesh.

This field stores the count of cells within the compartmental mesh.

_marker: PhantomData<T>

PhantomData marker for generic type T.

This field is used to mark the generic type T in the struct without actually storing data.

Implementations

impl<T> BaseCompartmentMesh<T>

fn new(descriptors: [AxisDescriptor; 3]) -> Self