cmtool_core::grid

Type Alias MeshCylindrical

pub type MeshCylindrical = BaseCompartmentMesh<CylindricalMarker>;

Cylindrical Compartment mesh structure for spatial modeling.

Aliased Type

pub struct MeshCylindrical {
    axes: [CoordAxis; 3],
    n_cells: usize,
    _marker: PhantomData<CylindricalMarker>,
}

Fields
axes: [CoordAxis; 3]

The axes of the coordinate system for the mesh.

This array contains three CoordAxis instances, each representing one of the principal axes

n_cells: usize

The total number of cells in the mesh.

This field stores the count of cells within the compartmental mesh.

_marker: PhantomData<CylindricalMarker>

PhantomData marker for generic type T.

This field is used to mark the generic type T in the struct without actually storing data.